# docs/swagger.yaml
basePath: /v1
definitions:
...
  attachment.insertContract:
    properties:
      amount_value:
        type: number
      comments:
        type: string
      contract_model:
        type: string
      description:
        type: string
      discount_value:
        type: number
      document:
        type: string
      due_date:
        type: string
      extra_amount_value:
        type: number
      installment_number:
        type: integer
      labels:
        items:
          type: string
        type: array
      person_id:
        type: string
      products:
        items:
          type: string
        type: array
      project_id:
        type: string
      reseller_id:
        type: string
      signed_at:
        type: string
      total_amount_value:
        type: number
    required:
    - amount_value
    - comments
    - contract_model
    - description
    - discount_value
    - due_date
    - extra_amount_value
    - total_amount_value
    type: object
  attachment.insertHelpDeskTicket:
    properties:
      code:
        type: integer
      due_date:
        type: string
      labels:
        items:
          type: string
        type: array
      level:
        type: string
      person_id:
        type: string
      project_id:
        type: string
      source:
        type: string
      source_id:
        type: string
      status:
        type: boolean
      title:
        type: string
    required:
    - code
    - due_date
    - level
    - source
    - source_id
    - title
    type: object
...
  attachment.insertLabel:
    properties:
      name:
        type: string
    required:
    - name
    type: object
...
  attachment.updateLabel:
    properties:
      name:
        type: string
    required:
    - name
    type: object
...
  attachment.updateProduct:
    properties:
      categories:
        items:
          type: string
        type: array
      description:
        type: string
      name:
        type: string
      prices:
        items:
          properties:
            code:
              type: string
            description:
              type: string
            id:
              type: string
            status:
              type: boolean
            title:
              type: string
            value:
              type: number
          required:
          - code
          - description
          - value
          type: object
        type: array
      status:
        type: boolean
    required:
    - description
    - name
    type: object
...
  entities.Contract:
    properties:
      amount_value:
        type: number
      comments:
        type: string
      contract_model:
        type: string
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      discount_value:
        type: number
      document:
        type: string
      due_date:
        type: string
      extra_amount_value:
        type: number
      id:
        type: string
      installment_number:
        type: integer
      invoices:
        items:
          $ref: '#/definitions/entities.Invoice'
        type: array
      labels:
        items:
          $ref: '#/definitions/entities.Label'
        type: array
      person:
        $ref: '#/definitions/entities.Person'
      person_id:
        type: string
      products:
        items:
          $ref: '#/definitions/entities.Product'
        type: array
      project:
        $ref: '#/definitions/entities.Project'
      project_id:
        type: string
      reseller_id:
        type: string
      signed_at:
        type: string
      total_amount_value:
        type: number
      updated_at:
        type: string
    required:
    - amount_value
    - comments
    - contract_model
    - description
    - discount_value
    - due_date
    - extra_amount_value
    - total_amount_value
    type: object
  entities.HelpDeskTicket:
    properties:
      code:
        type: integer
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      due_date:
        type: string
      id:
        type: string
      labels:
        items:
          $ref: '#/definitions/entities.Label'
        type: array
      level:
        type: string
      messages:
        items:
          $ref: '#/definitions/entities.HelpDeskTicketMessage'
        type: array
      person:
        $ref: '#/definitions/entities.Person'
      person_id:
        type: string
      project:
        $ref: '#/definitions/entities.Project'
      project_id:
        type: string
      source:
        type: string
      source_id:
        type: string
      status:
        type: boolean
      title:
        type: string
      updated_at:
        type: string
    required:
    - code
    - due_date
    - level
    - source
    - source_id
    - title
    type: object
  entities.HelpDeskTicketMessage:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      ticket_id:
        type: string
    required:
    - description
    - ticket_id
    type: object
  entities.Invoice:
    properties:
      amount_value:
        type: number
      bank_slip_code:
        type: string
      bank_slip_document:
        type: string
      bank_slip_input_code:
        type: string
      code:
        type: string
      comments:
        type: string
      contract_id:
        type: string
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      discount_comments:
        type: string
      discount_value:
        type: number
      due_date:
        type: string
      extra_amount_comments:
        type: string
      extra_amount_value:
        type: number
      id:
        type: string
      installment_number:
        type: integer
      items:
        items:
          $ref: '#/definitions/entities.InvoiceItem'
        type: array
      late_invoices:
        type: boolean
      paid_at:
        type: string
      payment_comments:
        type: string
      payment_method:
        type: string
      person:
        $ref: '#/definitions/entities.Person'
      person_id:
        type: string
      pix_copy_paste:
        type: string
      pix_qr_code:
        type: string
      status:
        type: boolean
      total_amount_value:
        type: number
      updated_at:
        type: string
      warning_comments:
        type: string
    required:
    - amount_value
    - code
    - discount_value
    - due_date
    - extra_amount_value
    - total_amount_value
    type: object
...
  entities.Label:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    required:
    - name
    type: object
  entities.Person:
    properties:
      addresses:
        items:
          $ref: '#/definitions/entities.PersonAddress'
        type: array
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      document:
        type: string
      emails:
        items:
          $ref: '#/definitions/entities.PersonEmail'
        type: array
      extras:
        items:
          $ref: '#/definitions/entities.PersonExtra'
        type: array
      id:
        type: string
      main_person_id:
        type: string
      name:
        type: string
      nickname:
        type: string
      phones:
        items:
          $ref: '#/definitions/entities.PersonPhone'
        type: array
      status:
        type: boolean
      updated_at:
        type: string
    required:
    - document
    - name
    type: object
...
  entities.PersonEmail:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      id:
        type: string
      is_main:
        type: boolean
      person_id:
        type: string
      updated_at:
        type: string
    required:
    - person_id
    type: object
...
  entities.Product:
    properties:
      categories:
        items:
          $ref: '#/definitions/entities.ProductCategory'
        type: array
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      id:
        type: string
      name:
        type: string
      prices:
        items:
          $ref: '#/definitions/entities.ProductPrice'
        type: array
      status:
        type: boolean
      updated_at:
        type: string
    required:
    - description
    - name
    type: object
  entities.ProductCategory:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      id:
        type: string
      name:
        type: string
      products:
        items:
          $ref: '#/definitions/entities.Product'
        type: array
      status:
        type: boolean
      updated_at:
        type: string
    required:
    - description
    - name
    type: object
...
  entities.Project:
    properties:
      comments:
        type: string
      content:
        type: string
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      deliveries:
        items:
          type: string
        type: array
      due_date:
        type: string
      id:
        type: string
      kind:
        type: string
      labels:
        items:
          $ref: '#/definitions/entities.Label'
        type: array
      person:
        $ref: '#/definitions/entities.Person'
      person_id:
        type: string
      repositories:
        items:
          $ref: '#/definitions/entities.Repositories'
        type: array
      requirements:
        items:
          type: string
        type: array
      title:
        type: string
      updated_at:
        type: string
    required:
    - content
    - kind
    - requirements
    - title
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  helpers.Error:
    properties:
      payload:
        properties:
          code:
            example: 500
            type: integer
          details:
            example: error details
            type: string
          message:
            example: error message
            type: string
        required:
        - code
        - details
        - message
        type: object
      success:
        example: false
        type: boolean
    required:
    - success
    type: object
  helpers.ResponseMany-array_entities_Contract:
    properties:
      pagination:
        $ref: '#/definitions/utils.Pagination'
      payload:
        items:
          $ref: '#/definitions/entities.Contract'
        type: array
      success:
        example: true
        type: boolean
    required:
    - success
    type: object
  helpers.ResponseMany-array_entities_HelpDeskTicket:
    properties:
      pagination:
        $ref: '#/definitions/utils.Pagination'
      payload:
        items:
          $ref: '#/definitions/entities.HelpDeskTicket'
        type: array
      success:
        example: true
        type: boolean
    required:
    - success
    type: object
...
  helpers.ResponseMany-array_entities_Label:
    properties:
      pagination:
        $ref: '#/definitions/utils.Pagination'
      payload:
        items:
          $ref: '#/definitions/entities.Label'
        type: array
      success:
        example: true
        type: boolean
    required:
    - success
    type: object
...
  helpers.ResponseNone:
    properties:
      success:
        example: true
        type: boolean
    required:
    - success
    type: object
  helpers.ResponseOne-entities_Contract:
    properties:
      payload:
        $ref: '#/definitions/entities.Contract'
      success:
        example: true
        type: boolean
    required:
    - success
    type: object
  helpers.ResponseOne-entities_HelpDeskTicket:
    properties:
      payload:
        $ref: '#/definitions/entities.HelpDeskTicket'
      success:
        example: true
        type: boolean
    required:
    - success
    type: object
...
  helpers.ResponseOne-entities_Label:
    properties:
      payload:
        $ref: '#/definitions/entities.Label'
      success:
        example: true
        type: boolean
    required:
    - success
    type: object
...
  helpers.ResponseOne-entities_Product:
    properties:
      payload:
        $ref: '#/definitions/entities.Product'
      success:
        example: true
        type: boolean
    required:
    - success
    type: object
...
  utils.Pagination:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      pages:
        items:
          type: integer
        type: array
      total_pages:
        type: integer
      total_registers:
        type: integer
    type: object
...
host: localhost:9000
info:
  contact: {}
  description: Welcome to the official documentation for the The Mug Codes manager.
    This guide provides comprehensive information to integrate and utilize our API
    effectively. It includes details on authentication, endpoints, rate limits, error
    handling, best practices, and support. Whether you're a beginner or an experienced
    developer, this documentation serves as your roadmap to leverage our services
    seamlessly.
  title: The Mug Codes
  version: "1.0"
paths:
...
  /contracts:
    get:
      description: Get all contracts from database.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.ResponseMany-array_entities_Contract'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.Error'
      summary: Show All Contract
      tags:
      - Contracts
    post:
      consumes:
      - application/json
      description: Creates a new contract in database.
      parameters:
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/attachment.insertContract'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/helpers.ResponseOne-entities_Contract'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.Error'
      summary: Create a Contract
      tags:
      - Contracts
...
  /helpdesk/tickets:
    get:
      description: Get all helpdesk tickets from database.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.ResponseMany-array_entities_HelpDeskTicket'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.Error'
      summary: Show All HelpDesk Tickets
      tags:
      - HelpDesk - Tickets
    post:
      consumes:
      - application/json
      description: Creates a new helpdesk ticket in database.
      parameters:
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/attachment.insertHelpDeskTicket'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/helpers.ResponseOne-entities_HelpDeskTicket'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.Error'
      summary: Create a HelpDesk Ticket
      tags:
      - HelpDesk - Tickets
...
  /invoices/pdf/{id}:
    get:
      description: Get a invoice pdf file by idin database.
      produces:
      - application/pdf
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.Error'
      summary: Show a Invoice PDF File
      tags:
      - Invoices
  /labels:
    get:
      description: Get all labels from database.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.ResponseMany-array_entities_Label'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.Error'
      summary: Show All Labels
      tags:
      - Labels
    post:
      consumes:
      - application/json
      description: Creates a new label in database.
      parameters:
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/attachment.insertLabel'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/helpers.ResponseOne-entities_Label'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.Error'
      summary: Create a Label
      tags:
      - Labels
  /labels/{id}:
    delete:
      description: Removes a label by idfrom database.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.ResponseNone'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.Error'
      summary: Delete a Label
      tags:
      - Labels
    get:
      description: Get a label by idin database.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.ResponseOne-entities_Label'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.Error'
      summary: Show a Label
      tags:
      - Labels
    put:
      consumes:
      - application/json
      description: Updates a label by idin database.
      parameters:
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/attachment.updateLabel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.ResponseOne-entities_Label'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.Error'
      summary: Change a Label
      tags:
      - Labels
...
  /products/{id}:
    delete:
      description: Removes a product by idfrom database.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.ResponseNone'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.Error'
      summary: Delete a Product
      tags:
      - Products
    get:
      description: Get a product by idin database.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.ResponseOne-entities_Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.Error'
      summary: Show a Products
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: Updates a product by idin database.
      parameters:
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/attachment.updateProduct'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.ResponseOne-entities_Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.Error'
      summary: Change a Product
      tags:
      - Products
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"